add Computers and do abstract class

+++ shop/src/Specific/VideoCard.java
@@ -6,7 +6,7 @@ public class VideoCard extends Cpu implements TechSpecification  {
     public VideoMemory videoCard;
 
 
-    public VideoCard(int core, float freq) {
+    public VideoCard(int core, double freq) {
         super(core, freq);
     }
 }
